feat(nvmx): abstract probe context
This shuffles some code around to abstract raw pointers
from the NVMe controller itself. The device creation itself
is moved to a separate file (uri.rs).
