+++ mayastor/src/bdev/dev/nvmx/controller.rs
@@ -1,178 +1,74 @@
-use async_trait::async_trait;
-use futures::channel::oneshot;
-use nix::errno::Errno;
-use snafu::ResultExt;
+//!
+//!
+//! This file contains the main structures for a NVMe controller
 use std::{
-    collections::HashMap,
-    convert::{From, TryFrom},
-    mem::size_of,
+    convert::From,
     os::raw::c_void,
-    ptr::{copy_nonoverlapping, NonNull},
+    ptr::NonNull,
     sync::{Arc, Mutex},
 };
-use tracing::instrument;
-use url::Url;
 
 use spdk_sys::{
-    self,
     spdk_io_device_register,
-    spdk_io_device_unregister,
-    spdk_nvme_connect_async,
     spdk_nvme_ctrlr,
-    spdk_nvme_ctrlr_get_default_ctrlr_opts,
     spdk_nvme_ctrlr_get_ns,
-    spdk_nvme_ctrlr_opts,
     spdk_nvme_ctrlr_process_admin_completions,
-    spdk_nvme_detach,
-    spdk_nvme_probe_ctx,
-    spdk_nvme_probe_poll_async,
-    spdk_nvme_transport_id,
     spdk_poller,
     spdk_poller_register_named,
-    spdk_poller_unregister,
 };
 
-use crate::{
-    bdev::{
-        dev::nvmx::{
-            channel::{NvmeControllerIoChannel, NvmeIoChannel},
-            NvmeNamespace,
-            NVME_CONTROLLERS,
-        },
-        util::uri,
-        CreateDestroy,
-        GetName,
-    },
-    ffihelper::ErrnoResult,
-    nexus_uri::{
-        NexusBdevError,
-        {self},
-    },
-    subsys::NvmeBdevOpts,
+use crate::bdev::dev::nvmx::{
+    channel::{NvmeControllerIoChannel, NvmeIoChannel},
+    nvme_bdev_running_config,
+    nvme_controller_lookup,
+    uri::NvmeControllerContext,
+    NvmeNamespace,
+    NVME_CONTROLLERS,
 };
-
-const DEFAULT_NVMF_PORT: u16 = 8420;
-
-#[derive(Debug)]
-pub struct NvmfDeviceTemplate {
-    /// name of the nvme controller and base name of the bdev
-    /// that should be created for each namespace found
-    name: String,
-    /// alias which can be used to open the bdev
-    alias: String,
-    /// the remote target host (address)
-    host: String,
-    /// the transport service id (ie. port)
-    port: u16,
-    /// the nqn of the subsystem we want to connect to
-    subnqn: String,
-    /// Enable protection information checking (reftag, guard)
-    prchk_flags: u32,
-    /// uuid of the spdk bdev
-    uuid: Option<uuid::Uuid>,
-}
-
-impl TryFrom<&Url> for NvmfDeviceTemplate {
-    type Error = NexusBdevError;
-
-    fn try_from(url: &Url) -> Result<Self, Self::Error> {
-        let host =
-            url.host_str().ok_or_else(|| NexusBdevError::UriInvalid {
-                uri: url.to_string(),
-                message: String::from("missing host"),
-            })?;
-
-        let segments = uri::segments(url);
-
-        if segments.is_empty() {
-            return Err(NexusBdevError::UriInvalid {
-                uri: url.to_string(),
-                message: String::from("no path segment"),
-            });
-        }
-
-        if segments.len() > 1 {
-            return Err(NexusBdevError::UriInvalid {
-                uri: url.to_string(),
-                message: String::from("too many path segments"),
-            });
-        }
-
-        let mut parameters: HashMap<String, String> =
-            url.query_pairs().into_owned().collect();
-
-        let mut prchk_flags: u32 = 0;
-
-        if let Some(value) = parameters.remove("reftag") {
-            if uri::boolean(&value, true).context(
-                nexus_uri::BoolParamParseError {
-                    uri: url.to_string(),
-                    parameter: String::from("reftag"),
-                },
-            )? {
-                prchk_flags |= spdk_sys::SPDK_NVME_IO_FLAGS_PRCHK_REFTAG;
-            }
-        }
-
-        if let Some(value) = parameters.remove("guard") {
-            if uri::boolean(&value, true).context(
-                nexus_uri::BoolParamParseError {
-                    uri: url.to_string(),
-                    parameter: String::from("guard"),
-                },
-            )? {
-                prchk_flags |= spdk_sys::SPDK_NVME_IO_FLAGS_PRCHK_GUARD;
-            }
-        }
-
-        let uuid = uri::uuid(parameters.remove("uuid")).context(
-            nexus_uri::UuidParamParseError {
-                uri: url.to_string(),
-            },
-        )?;
-
-        if let Some(keys) = uri::keys(parameters) {
-            warn!("ignored parameters: {}", keys);
-        }
-
-        Ok(NvmfDeviceTemplate {
-            name: url[url::Position::BeforeHost .. url::Position::AfterPath]
-                .to_string(),
-            alias: url.to_string(),
-            host: host.to_string(),
-            port: url.port().unwrap_or(DEFAULT_NVMF_PORT),
-            subnqn: segments[0].to_string(),
-            prchk_flags,
-            uuid,
-        })
-    }
-}
-
-impl GetName for NvmfDeviceTemplate {
-    fn get_name(&self) -> String {
-        format!("{}n1", self.name)
-    }
-}
 
 #[derive(Debug, PartialEq)]
 pub enum NvmeControllerState {
     Initializing,
     Running,
 }
-pub struct NvmeControllerInner {
-    namespaces: Vec<Arc<NvmeNamespace>>,
-    ctrlr: NonNull<spdk_nvme_ctrlr>,
-    adminq_poller: NonNull<spdk_poller>,
+
+#[derive(Debug)]
+pub(crate) struct NvmeControllerInner {
+    pub(crate) namespaces: Vec<Arc<NvmeNamespace>>,
+    pub(crate) ctrlr: NonNull<spdk_nvme_ctrlr>,
+    pub(crate) adminq_poller: NonNull<spdk_poller>,
 }
+
+impl NvmeControllerInner {
+    fn new(ctrlr: NonNull<spdk_nvme_ctrlr>) -> Self {
+        let adminq_poller = NonNull::new(unsafe {
+            spdk_poller_register_named(
+                Some(nvme_poll_adminq),
+                ctrlr.as_ptr().cast(),
+                nvme_bdev_running_config().nvme_adminq_poll_period_us,
+                "nvme_poll_adminq\0" as *const _ as *mut _,
+            )
+        })
+        .expect("failed to create poller");
+
+        Self {
+            ctrlr,
+            adminq_poller,
+            namespaces: Vec::new(),
+        }
+    }
+}
+
 /*
  * NVME controller implementation.
  */
+#[derive(Debug)]
 pub struct NvmeController {
-    name: String,
-    id: u64,
-    prchk_flags: u32,
-    state: NvmeControllerState,
-    inner: Option<NvmeControllerInner>,
+    pub(crate) name: String,
+    pub(crate) id: u64,
+    pub(crate) prchk_flags: u32,
+    pub(crate) state: NvmeControllerState,
+    pub(crate) inner: Option<NvmeControllerInner>,
 }
 
 unsafe impl Send for NvmeController {}
@@ -185,7 +81,8 @@ impl Drop for NvmeController {
 }
 
 impl NvmeController {
-    fn new(name: &str, prchk_flags: u32) -> Self {
+    /// Creates a new NVMe controller with the given name.
+    pub fn new(name: &str, prchk_flags: u32) -> Option<Self> {
         let l = NvmeController {
             name: String::from(name),
             id: 0,
@@ -194,18 +91,21 @@ impl NvmeController {
             inner: None,
         };
 
-        debug!("{}: New controller created", l.get_name());
-        l
+        debug!("{}: new NVMe controller created", l.get_name());
+        Some(l)
     }
 
+    /// returns the name of the current controller
     pub fn get_name(&self) -> String {
         self.name.clone()
     }
 
-    pub fn get_flags(&self) -> u32 {
+    /// returns the protection flags the controller is created with
+    pub fn flags(&self) -> u32 {
         self.prchk_flags
     }
 
+    /// returns the ID of the controller
     pub fn id(&self) -> u64 {
         assert_ne!(self.id, 0, "Controller ID is not yet initialized");
         self.id
@@ -220,40 +120,66 @@ impl NvmeController {
 
     // As of now, only 1 namespace per controller is supported.
     pub fn namespace(&self) -> Option<Arc<NvmeNamespace>> {
-        if self.inner.is_none() {
+        let inner = self
+            .inner
+            .as_ref()
+            .expect("(BUG) no inner NVMe controller defined yet");
+
+        if let Some(ns) = inner.namespaces.get(0) {
+            Some(Arc::clone(ns))
+        } else {
+            debug!("no namespaces associated with the current controller");
             None
-        } else {
-            let inner = self.inner.as_ref().unwrap();
-            if let Some(ns) = inner.namespaces.get(0) {
-                Some(Arc::clone(ns))
-            } else {
-                None
-            }
         }
     }
 
-    pub fn spdk_handle(&self) -> *mut spdk_nvme_ctrlr {
-        if self.inner.is_none() {
-            error!("{} No SPDK handle configured !", self.name);
-            std::ptr::null_mut()
-        } else {
-            debug!(
-                "{} SPDK handle: {:p}",
-                self.name,
-                self.inner.as_ref().unwrap().ctrlr.as_ptr()
+    /// register the controller as an io device
+    fn register_io_device(&self) {
+        unsafe {
+            spdk_io_device_register(
+                self.id() as *mut c_void,
+                Some(NvmeControllerIoChannel::create),
+                Some(NvmeControllerIoChannel::destroy),
+                std::mem::size_of::<NvmeIoChannel>() as u32,
+                self.get_name().as_ptr() as *const i8,
+            )
+        }
+
+        debug!(
+            "{}: I/O device registered at 0x{:X}",
+            self.get_name(),
+            self.id()
+        );
+    }
+
+    /// we should try to avoid this
+    pub fn ctrlr_as_ptr(&self) -> *mut spdk_nvme_ctrlr {
+        self.inner.as_ref().map_or(std::ptr::null_mut(), |c| {
+            let ptr = c.ctrlr.as_ptr();
+            debug!("SPDK handle {:p}", ptr);
+            ptr
+        })
+    }
+
+    /// populate name spaces, at current we only populate the first namespace
+    fn populate_namespaces(&mut self) {
+        let ns = unsafe { spdk_nvme_ctrlr_get_ns(self.ctrlr_as_ptr(), 1) };
+
+        if ns.is_null() {
+            warn!(
+                "{} no namespaces reported by the NVMe controller",
+                self.get_name()
             );
-            self.inner.as_ref().unwrap().ctrlr.as_ptr()
         }
+
+        self.inner.as_mut().unwrap().namespaces =
+            vec![Arc::new(NvmeNamespace::from_ptr(ns))]
     }
 }
 
-extern "C" fn nvme_async_poll(ctx: *mut c_void) -> i32 {
-    let _rc =
-        unsafe { spdk_nvme_probe_poll_async(ctx as *mut spdk_nvme_probe_ctx) };
-    1
-}
-
-extern "C" fn nvme_poll_adminq(ctx: *mut c_void) -> i32 {
+/// return number of completions processed (maybe 0) or negated on error. -ENXIO
+//  in the special case that the qpair is failed at the transport layer.
+pub extern "C" fn nvme_poll_adminq(ctx: *mut c_void) -> i32 {
     //println!("adminq poll");
 
     let rc = unsafe {
@@ -267,295 +193,45 @@ extern "C" fn nvme_poll_adminq(ctx: *mut c_void) -> i32 {
     }
 }
 
-// Callback to be called once NVME controller successfully created.
-extern "C" fn connect_attach_cb(
-    _cb_ctx: *mut c_void,
-    _trid: *const spdk_nvme_transport_id,
-    ctrlr: *mut spdk_nvme_ctrlr,
-    _opts: *const spdk_nvme_ctrlr_opts,
+pub(crate) fn connected_attached_cb(
+    ctx: &mut NvmeControllerContext,
+    ctrlr: NonNull<spdk_nvme_ctrlr>,
 ) {
-    let context =
-        unsafe { &mut *(_cb_ctx as *const _ as *mut NvmeControllerContext) };
+    ctx.unregister_poller();
+    // we use the ctrlr address as the controller id in the global table
+    let cid = ctrlr.as_ptr() as u64;
 
-    let mut poller = context.poller.take().unwrap();
-    unsafe {
-        spdk_poller_unregister(&mut poller);
-    }
+    // get a reference to our controller we created when we kicked of the async
+    // attaching process
+    let controller =
+        nvme_controller_lookup(&ctx.name()).expect("no controller in the list");
 
-    let cid = ctrlr as u64;
+    // clone it now such that we can lock the original, and insert it later.
+    let ctl = Arc::clone(&controller);
 
-    // Save SPDK controller reference for further use.
-    let controllers = NVME_CONTROLLERS.write().unwrap();
-    // Create a clone of controller to insert later for ID lookup.
-    let rc = controllers.get(&context.name).unwrap();
-    let clone = Arc::clone(rc);
-    let mut controller = rc.lock().unwrap();
+    let mut controller = controller.lock().unwrap();
 
     controller.set_id(cid);
-    // Register I/O device for the controller.
-    unsafe {
-        spdk_io_device_register(
-            controller.id() as *mut c_void,
-            Some(NvmeControllerIoChannel::create),
-            Some(NvmeControllerIoChannel::destroy),
-            std::mem::size_of::<NvmeIoChannel>() as u32,
-            controller.get_name().as_ptr() as *const i8,
-        )
-    }
+    controller.inner = Some(NvmeControllerInner::new(ctrlr));
+    controller.register_io_device();
+
     debug!(
         "{}: I/O device registered at 0x{:X}",
         controller.get_name(),
         controller.id()
     );
 
-    // Configure poller for controller's admin queue.
-    let default_opts = NvmeBdevOpts::default();
-
-    let adminq_poller = unsafe {
-        spdk_poller_register_named(
-            Some(nvme_poll_adminq),
-            ctrlr as *mut c_void,
-            default_opts.nvme_adminq_poll_period_us,
-            "nvme_poll_adminq\0" as *const _ as *mut _,
-        )
-    };
-
-    if adminq_poller.is_null() {
-        error!(
-            "{}: failed to create admin queue poller",
-            controller.get_name()
-        );
-    }
-
-    let mut namespaces: Vec<Arc<NvmeNamespace>> = Vec::new();
-    // Initialize namespaces (currently only 1).
-    let ns = unsafe { spdk_nvme_ctrlr_get_ns(ctrlr, 1) };
-
-    if ns.is_null() {
-        warn!(
-            "{} no namespaces reported by the NVMe controller",
-            controller.get_name()
-        );
-    } else {
-        namespaces.push(Arc::new(NvmeNamespace::from_ptr(ns)));
-    }
-
-    let inner_state = NvmeControllerInner {
-        ctrlr: NonNull::new(ctrlr).unwrap(),
-        adminq_poller: NonNull::new(adminq_poller).unwrap(),
-        namespaces,
-    };
-
-    controller.inner.replace(inner_state);
+    controller.populate_namespaces();
     controller.state = NvmeControllerState::Running;
 
-    // Release the guard early to let the waiter access the controller instance
-    // safely.
-    drop(controller);
-    drop(controllers);
-
-    // Add 'controller id -> controller' mapping once the controller is fully
-    // setup.
-    let mut controllers = NVME_CONTROLLERS.write().unwrap();
-    controllers.insert(cid.to_string(), clone);
-
+    nvme_controller_insert(cid, ctl);
     // Wake up the waiter and complete controller registration.
-    let sender = context.sender.take().unwrap();
-    sender
+    ctx.sender()
         .send(Ok(()))
         .expect("done callback receiver side disappeared");
 }
 
-#[async_trait(?Send)]
-impl CreateDestroy for NvmfDeviceTemplate {
-    type Error = NexusBdevError;
-
-    #[instrument(err)]
-    async fn create(&self) -> Result<String, Self::Error> {
-        let cname = self.get_name();
-
-        // Check against existing controller with the transport ID.
-        let mut controllers = NVME_CONTROLLERS.write().unwrap();
-        if controllers.contains_key(&cname) {
-            return Err(NexusBdevError::BdevExists {
-                name: cname,
-            });
-        }
-
-        // Insert a new controller instance (uninitialized) as a guard, and
-        // release the lock to keep the write path as short, as
-        // possible.
-        let rc =
-            Arc::new(Mutex::new(NvmeController::new(&cname, self.prchk_flags)));
-        controllers.insert(cname.clone(), rc);
-        drop(controllers);
-
-        let mut context = NvmeControllerContext::new(self);
-
-        // Initiate connection with remote NVMe target.
-        let probe_ctx: *mut spdk_nvme_probe_ctx = unsafe {
-            spdk_nvme_connect_async(
-                &context.trid,
-                &context.opts,
-                Some(connect_attach_cb),
-            )
-        };
-
-        if probe_ctx.is_null() {
-            // Remove controller record before returning error.
-            let mut controllers = NVME_CONTROLLERS.write().unwrap();
-            controllers.remove(&cname);
-
-            return Err(NexusBdevError::CreateBdev {
-                name: cname,
-                source: Errno::ENODEV,
-            });
-        }
-
-        // Register poller to check for connection status.
-        let poller = unsafe {
-            spdk_poller_register_named(
-                Some(nvme_async_poll),
-                probe_ctx as *mut c_void,
-                1000, // TODO: fix.
-                "nvme_async_poll\0" as *const _ as *mut _,
-            )
-        };
-
-        context.poller = Some(poller);
-
-        context
-            .receiver
-            .await
-            .context(nexus_uri::CancelBdev {
-                name: self.name.clone(),
-            })?
-            .context(nexus_uri::CreateBdev {
-                name: self.name.clone(),
-            })?;
-
-        // Check that controller is fully initialized.
-        let controllers = NVME_CONTROLLERS.read().unwrap();
-        let controller = controllers.get(&cname).unwrap().lock().unwrap();
-        assert_eq!(
-            controller.state,
-            NvmeControllerState::Running,
-            "NVMe controller is not fully initialized"
-        );
-
-        info!("{} NVMe controller successfully initialized", cname);
-        Ok(cname)
-    }
-
-    // nvme_bdev_ctrlr_create
-    async fn destroy(self: Box<Self>) -> Result<(), Self::Error> {
-        let cname = self.get_name();
-
-        let mut controllers = NVME_CONTROLLERS.write().unwrap();
-        if !controllers.contains_key(&cname) {
-            return Err(NexusBdevError::BdevNotFound {
-                name: cname,
-            });
-        }
-
-        // Remove 'controller name -> controller' mapping.
-        let e = controllers.remove(&cname).unwrap();
-        let mut controller = e.lock().unwrap();
-        debug!("{}: removing NVMe controller", cname);
-
-        if let Some(inner) = controller.inner.take() {
-            debug!("{} unregistering adminq poller", controller.get_name());
-            unsafe {
-                spdk_poller_unregister(&mut inner.adminq_poller.as_ptr());
-            }
-
-            debug!(
-                "{}: unregistering I/O device at 0x{:X}",
-                controller.get_name(),
-                controller.id()
-            );
-            unsafe {
-                spdk_io_device_unregister(controller.id() as *mut c_void, None);
-            }
-
-            // Detach SPDK controller.
-            let rc = unsafe { spdk_nvme_detach(inner.ctrlr.as_ptr()) };
-
-            assert_eq!(rc, 0, "Failed to detach NVMe controller");
-            debug!("{}: NVMe controller successfully detached", cname);
-        }
-
-        // Remove 'controller id->controller' mappig.
-        controllers.remove(&controller.id().to_string());
-
-        Ok(())
-    }
-}
-
-// Context for an NVMe controller being created.
-struct NvmeControllerContext {
-    opts: spdk_nvme_ctrlr_opts,
-    name: String,
-    trid: spdk_nvme_transport_id,
-    sender: Option<oneshot::Sender<Result<(), Errno>>>,
-    receiver: oneshot::Receiver<Result<(), Errno>>,
-    poller: Option<*mut spdk_poller>,
-}
-
-impl NvmeControllerContext {
-    pub fn new(template: &NvmfDeviceTemplate) -> NvmeControllerContext {
-        let port = template.port.to_string();
-        let protocol = "tcp";
-
-        let default_opts = NvmeBdevOpts::default();
-        let mut trid = spdk_nvme_transport_id::default();
-        let mut opts = spdk_nvme_ctrlr_opts::default();
-
-        // Initialize options for NVMe controller to defaults.
-        unsafe {
-            spdk_nvme_ctrlr_get_default_ctrlr_opts(
-                &mut opts,
-                size_of::<spdk_nvme_ctrlr_opts>() as u64,
-            );
-        }
-
-        opts.transport_retry_count = default_opts.retry_count as u8;
-
-        unsafe {
-            copy_nonoverlapping(
-                protocol.as_ptr() as *const c_void,
-                &mut trid.trstring[0] as *const _ as *mut c_void,
-                protocol.len(),
-            );
-            copy_nonoverlapping(
-                template.host.as_ptr() as *const c_void,
-                &mut trid.traddr[0] as *const _ as *mut c_void,
-                template.host.len(),
-            );
-            copy_nonoverlapping(
-                port.as_ptr() as *const c_void,
-                &mut trid.trsvcid[0] as *const _ as *mut c_void,
-                port.len(),
-            );
-            copy_nonoverlapping(
-                template.subnqn.as_ptr() as *const c_void,
-                &mut trid.subnqn[0] as *const _ as *mut c_void,
-                template.subnqn.len(),
-            );
-        }
-
-        trid.trtype = spdk_sys::SPDK_NVME_TRANSPORT_TCP;
-        trid.adrfam = spdk_sys::SPDK_NVMF_ADRFAM_IPV4;
-
-        let (sender, receiver) = oneshot::channel::<ErrnoResult<()>>();
-
-        NvmeControllerContext {
-            opts,
-            trid,
-            name: template.get_name(),
-            sender: Some(sender),
-            receiver,
-            poller: None,
-        }
-    }
+fn nvme_controller_insert(cid: u64, ctl: Arc<Mutex<NvmeController>>) {
+    let mut controllers = NVME_CONTROLLERS.write().unwrap();
+    controllers.insert(cid.to_string(), ctl);
 }
